settings: Move confirm_reactivation out of settings_users.js.
This better organization may also simplify removing import cycles in
the future.
